+++ tests/_models/torchvision/test_convnext.py
@@ -8,30 +8,30 @@
 import torch
 from torchvision import models
 
-from lightly_train._models.torchvision.convnext import ConvNeXtFeatureExtractor
+from lightly_train._models.torchvision.convnext import ConvNeXtModelWrapper
 
 
-class TestConvNeXtFeatureExtractor:
+class TestConvNeXtModelWrapper:
     def test_feature_dim(self) -> None:
         model = models.convnext_tiny()
-        feature_extractor = ConvNeXtFeatureExtractor(model=model)
+        feature_extractor = ConvNeXtModelWrapper(model=model)
         assert feature_extractor.feature_dim() == 768
 
     def test_forward_features(self) -> None:
         model = models.convnext_tiny()
-        feature_extractor = ConvNeXtFeatureExtractor(model=model)
+        feature_extractor = ConvNeXtModelWrapper(model=model)
         x = torch.rand(1, 3, 224, 224)
         features = feature_extractor.forward_features(x)["features"]
         assert features.shape == (1, 768, 7, 7)
 
     def test_forward_pool(self) -> None:
         model = models.convnext_tiny()
-        feature_extractor = ConvNeXtFeatureExtractor(model=model)
+        feature_extractor = ConvNeXtModelWrapper(model=model)
         x = torch.rand(1, 768, 7, 7)
         pool = feature_extractor.forward_pool({"features": x})["pooled_features"]
         assert pool.shape == (1, 768, 1, 1)
 
     def test_get_model(self) -> None:
         model = models.convnext_tiny()
-        feature_extractor = ConvNeXtFeatureExtractor(model=model)
+        feature_extractor = ConvNeXtModelWrapper(model=model)
         assert feature_extractor.get_model() is model